Version 2 of the code
in which the number of threads is variable. At the moment the program has only two threads but what is desired now is that the user not only passes as argument the value of max but also the number of threads that are wanted to be executed in the program.